# PERF: construction with integer dtype