Motivated by this arithmetic example:

```
In [2]: ser = pd.Series(range(10_000_000))

In [3]: %timeit ser + 1
23.3 ms ± 304 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)  # <- main
15.1 ms ± 110 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)  # <- PR
```