@jbrockmendel 2022-12-17
> mypy complains.

fixed

> So copy wasn't used anyway?

correct